+++ finagle-core/src/main/scala/com/twitter/finagle/socks/flags.scala
@@ -8,6 +8,10 @@ private[finagle] object socksProxyHost extends GlobalFlag("", "SOCKS proxy host"
 private[finagle] object socksProxyPort extends GlobalFlag(0, "SOCKS proxy port") {
   override val name = "socksProxyPort"
 }
+private[finagle] object socksProxyForLocalhost
+    extends GlobalFlag(false, "Use SOCKS proxy for localhost connections, defaults to false.") {
+  override val name = "socksProxyForLocalhost"
+}
 private[finagle] object socksUsernameFlag extends GlobalFlag("", "SOCKS username") {
   override val name = "socksUsername"
 }